+++ src/perm/factored_traversal.rs
